Corrige la consistencia de estilo y formato del código
Se ajustaron espacios, comentarios y alineación para mantener un formato consistente en todo el proyecto. Esto facilita la lectura y el mantenimiento del código sin alterar la funcionalidad existente.

Se arreglo un bug de poder registrar Usuarios sin que se haya obtenido informacion del player anteriormente

+++ backend/app/models/TempPhrase/crud.py
@@ -1,9 +1,13 @@
-from httpx import AsyncClient
-from fastapi import HTTPException, status
-from faker import Faker
-from sqlalchemy.orm import Session
-from .model import TempPhrase
 from datetime import datetime, timedelta, timezone
+
+from faker import Faker
+from fastapi import HTTPException, status
+from httpx import AsyncClient
+from sqlalchemy.orm import Session
+
+from .model import TempPhrase
+
+from ..Player.model import Player
 
 async def get_player_description(player_id: int) -> str:
     url = f"https://users.roblox.com/v1/users/{player_id}"
@@ -19,13 +23,21 @@ async def get_player_description(player_id: int) -> str:
 
     return response["description"]
 
+
 fake = Faker("en_US")
+
 
 def create_random_phrase(quantity=20) -> str:
     return " ".join(fake.words(nb=quantity))
 
-def create_temp_phrase(db:Session, player_id) -> str:
+
+def create_temp_phrase(db: Session, player_id) -> str:
     existing_temp_phrase = db.query(TempPhrase).filter(TempPhrase.id_player == player_id).first()
+    existing_player = db.query(Player).filter(Player.id == player_id).first()
+
+    if not existing_player:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Player not found")
+
     if existing_temp_phrase:
         db.delete(existing_temp_phrase)
         db.commit()
@@ -33,7 +45,7 @@ def create_temp_phrase(db:Session, player_id) -> str:
     temp_phrase = TempPhrase(
         id_player=player_id,
         phrase=create_random_phrase(),
-        exp= datetime.now(timezone.utc) + timedelta(minutes=15)
+        exp=datetime.now(timezone.utc) + timedelta(minutes=15)
     )
     db.add(temp_phrase)
     db.commit()
